requirements.txt
fastapi
uvicorn
pathway
python-dotenv
pydantic
requests
mistralai
sentence-transformers
